Split the check connection workflow into multiple files.
