+++ airbyte-workers/src/main/java/io/airbyte/workers/temporal/check/connection/CheckConnectionWorkflow.java
@@ -1,0 +1,17 @@
+package io.airbyte.workers.temporal.check.connection;
+
+import io.airbyte.config.StandardCheckConnectionInput;
+import io.airbyte.config.StandardCheckConnectionOutput;
+import io.airbyte.scheduler.models.IntegrationLauncherConfig;
+import io.airbyte.scheduler.models.JobRunConfig;
+import io.temporal.workflow.WorkflowInterface;
+import io.temporal.workflow.WorkflowMethod;
+
+@WorkflowInterface
+public interface CheckConnectionWorkflow {
+
+  @WorkflowMethod
+  StandardCheckConnectionOutput run(JobRunConfig jobRunConfig,
+                                    IntegrationLauncherConfig launcherConfig,
+                                    StandardCheckConnectionInput connectionConfiguration);
+}
